lessEscapePlugin created
